bot/requirements.txt
aiogram>=3.0.0
pymongo
requests
python-dotenv
